Require callers to ChannelManager::AddChannel() to assign IDs.
This is so that we'll be able to use "process identifiers" as IDs, which
will be fine as long as we only have at most single channel (managed by
the same manager) between any given pair of processes. (If we change
that assumption/limitation, we'll have to munge process identifiers
somehow.)

R=yzshen@chromium.org

Review URL: https://codereview.chromium.org/887813002

+++ mojo/edk/embedder/embedder_internal.h
@@ -19,7 +19,7 @@ class ChannelManager;
 class Core;
 
 // Repeat a typedef in mojo/edk/system/channel_manager.h, to avoid including it.
-typedef uintptr_t ChannelId;
+typedef uint64_t ChannelId;
 
 }  // namespace system
